Moved PasswordForm in separate widget
