--- packages/client/src/pages/ProfilePage/ui/ChangePasswordForm/index.ts
@@ -1,1 +1,0 @@
-export { ChangePasswordForm } from './ChangePasswordForm';
