+++ packages/client/src/widgets/PasswordForm/index.ts
@@ -1,0 +1,1 @@
+export { PasswordForm } from './ui';
