+++ packages/client/src/widgets/PasswordForm/ui/PasswordForm/PasswordForm.interfaces.tsx
@@ -1,4 +1,4 @@
-export interface IChangePasswordFormProps {
+export interface IPasswordFormProps {
   isPasswordModalVisible: boolean;
   setIsPasswordModalVisible: (visible: boolean) => void;
 }
