+++ packages/client/src/widgets/PasswordForm/ui/index.ts
@@ -1,0 +1,1 @@
+export { PasswordForm } from './PasswordForm/PasswordForm';
